+++ src/enums/mine-types.enum.js
@@ -1,0 +1,32 @@
+// Docs: https://docs.w3cub.com/http/basics_of_http/mime_types/complete_list_of_mime_types.html
+
+export const MINE_TYPES = [
+  {
+    file_type: 'image/jpeg',
+    extension: ['.jpeg', '.jpg'],
+  },
+  {
+    file_type: 'image/png',
+    extension: ['.png'],
+  },
+  {
+    file_type: 'image/webp',
+    extension: ['.webp'],
+  },
+  {
+    file_type: 'image/svg+xml',
+    extension: ['.svg'],
+  },
+  {
+    file_type: 'image/gif',
+    extension: ['.gif'],
+  },
+  {
+    file_type: 'audio/mpeg',
+    extension: ['.mp3'],
+  },
+  {
+    file_type: 'video/mpeg',
+    extension: ['.mpeg'],
+  },
+]