+++ src/languages/vi/common.js
@@ -33,10 +33,8 @@ export default {
   parameter_wrong: 'YOUR PARAMETER PASSED WRONG!!!',
   editor_placeholder: 'Vui lòng nhập chữ',
   limit_size: 'Vui lòng chỉ tải tệp ảnh dưới {size}MB',
-  upload_image_error:
-    'エラーが発生しましたため、画像アップロードできませんでした',
-  rules_type_image:
-    '使用できない画像形式が含まれています。{type}のいずれかの画像をご使用ください',
+  upload_image_error: 'エラーが発生しましたため、画像アップロードできませんでした',
+  rules_type_image: '使用できない画像形式が含まれています。{type}のいずれかの画像をご使用ください',
   type_tag_script_err_mess: 'Illegal error (script) occurred',
   action_fail: 'Thao tác của bạn không thành công!',
   create_message_successfully: 'Tạo mới thành công!',
@@ -51,6 +49,7 @@ export default {
   popcomfirm_accept_btn: 'Xác nhận',
   popcomfirm_cancel_btn: 'Huỷ',
   please_enter: 'Vui lòng nhập',
+  please_select_format_type: 'Vui lòng chọn file có định dạng {type}',
   record: 'mục',
   total_number: 'Tổng cộng',
   strip_parameter_wrong: