+++ CHANGELOG.md
@@ -17,7 +17,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Multiple languages.
 - Authentication.
 
-
 ## [1.0.1] - 2022-12-20
 ### Added
 - Build CRUD feature for events.
@@ -25,11 +24,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Complete form components such as: InputSelect, InputTextNumber, InputUpload,...
 - Switch languages.
 - Function to delete stored data when leaving the place where they need to be used.
-
 ### Security
 - Set properties (secure, sameSite) to protect cookies.
 - Remove process login api in network tab when login done.
 - Block XSS for all input fields.
+- 
+## [1.0.2] - 2023-02-02
+### Fixed
+- Fix error messages.
+- Double-check the file type more deeply for InputUpload component.
 
 
 ## --- [Released] ---